backend/pyproject.toml
[project]
name = "aregrp-backend"
version = "0.1.0"
description = "Backend API (Django + Ninja)"
requires-python = ">=3.12"
dependencies = [
    "dj-database-url>=2.3.0",
    "django>=5.2.1",
    "django-cors-headers>=4.7.0",
    "django-ninja>=1.1.0",
    "django-storages>=1.14.2",
    "pre-commit>=4.2.0",
    "pyjwt>=2.8.0",
    "python-decouple>=3.8",
    "pyyaml>=6.0.2",
    "ruff>=0.11.13",
    "uvicorn>=0.34.2",
    "psycopg2-binary>=2.9.11",
    "boto3>=1.34.0",
    "pillow>=10.0.0",
    "django-smart-selects>=1.6.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-django>=4.8.0",
    "pytest-asyncio>=0.23.0",
]
